requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.25
alembic>=1.13.0
Jinja2>=3.1.4
python-multipart>=0.0.9
requests>=2.32.0
recipe-scrapers>=14.56.0
pydantic>=2.5.0
pytest>=7.4.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
